+++ tests/easy_uds_session_benchmark.cpp
@@ -1,7 +1,7 @@
 #include "easy_uds/easy_uds.hpp"
 
 #ifdef EASY_UDS_TRACE_SESSION_CONTENTION
-#include "../src/session_trace.hpp"
+#include "../src/client/trace.hpp"
 #endif
 
 #include <atomic>